+++ src/components/Map/OverlayStyles.js
@@ -1,0 +1,24 @@
+const overlayStyles = {
+  geolocateStyle: {
+    top: 100,
+    left: 0,
+    padding: "10px",
+  },
+  fullscreenControlStyle: {
+    top: 140,
+    left: 0,
+    padding: "10px",
+  },
+  navStyle: {
+    top: 180,
+    left: 0,
+    padding: "10px",
+  },
+  scaleControlStyle: {
+    bottom: 36,
+    left: 0,
+    padding: "10px",
+  },
+};
+
+export default overlayStyles;
